early stopping when accuracy over 80%

+++ src/build_network.h
@@ -1,10 +1,10 @@
 typedef struct{
   char *tag;
-  float value;
-  float bias;
-  float error;
-  float *weights;
-  float weighted_input;
+  double value;
+  double bias;
+  double error;
+  double *weights;
+  double weighted_input;
 }Node;
 
 int initialize_input_layer(char *input_value_tags[], Node *input_layer, int n);
